+++ R/frNN.R
@@ -65,6 +65,7 @@
 #' \item{dist }{a list with distances (same structure as
 #' `id`). }
 #' \item{eps }{ neighborhood radius `eps` that was used. }
+#' \item{metric }{ used distance metric. }
 #'
 #' `adjacencylist()` returns a list with one entry per data point in `x`. Each entry
 #' contains the id of the nearest neighbors.
@@ -81,6 +82,7 @@
 #'
 #' # Example 1: Find fixed radius nearest neighbors for each point
 #' nn <- frNN(x, eps = .5)
+#' nn
 #'
 #' # Number of neighbors
 #' hist(lengths(adjacencylist(nn)),
@@ -195,6 +197,7 @@ frNN <-
         )
       names(ret$dist) <- rownames(query)
       names(ret$id) <- rownames(query)
+      ret$metric = "euclidean"
     } else{
       ret <- frNN_int(
         as.matrix(x),
@@ -206,6 +209,7 @@ frNN <-
       )
       names(ret$dist) <- rownames(x)
       names(ret$id) <- rownames(x)
+      ret$metric = "euclidean"
     }
 
     ret$eps <- eps
@@ -261,6 +265,7 @@ dist_to_frNN <- function(x, eps, sort = FALSE) {
       dist = d,
       id = id,
       eps = eps,
+      metric = attr(x, "method"),
       sort = FALSE
     ),
       class = c("frNN", "NN"))
@@ -325,5 +330,7 @@ print.frNN <- function(x, ...) {
     "\n",
     sep = ""
   )
-  cat("Available fields: ", toString(names(x)), "\n", sep = "")
+
+  cat("Distance metric:", x$metric, "\n")
+  cat("\nAvailable fields: ", toString(names(x)), "\n", sep = "")
 }