+++ R/sNN.R
@@ -22,17 +22,24 @@
 
 #' Find Shared Nearest Neighbors
 #'
-#' Calculates the number of shared nearest neighbors, the shared nearest
-#' neighbor similarity and creates a shared nearest neighbors graph.
+#' Calculates the number of shared nearest neighbors
+#' and creates a shared nearest neighbors graph.
 #'
-#' The number of shared nearest neighbors is the intersection of the kNN
-#' neighborhood of two points. Note: that each point is considered to be part
-#' of its own kNN neighborhood. The range for the shared nearest neighbors is
-#' \eqn{[0, k]}.
+#' The number of shared nearest neighbors of two points p and q is the
+#' intersection of the kNN neighborhood of two points.
+#' Note: that each point is considered to be part
+#' of its own kNN neighborhood.
+#' The range for the shared nearest neighbors is
+#' \eqn{[0, k]}. The result is a n-by-k matrix called `shared`.
+#' Each row is a point and the columns are the point's k nearest neighbors.
+#' The value is the count of the shared neighbors.
 #'
-#' Javis and Patrick (1973) use the shared nearest neighbor graph for
-#' clustering. They only count shared neighbors between points that are in each
-#' other's kNN neighborhood.
+#' The shared nearest neighbor graph connects a point with all its nearest neighbors
+#' if they have at least one shared neighbor. The number of shared neighbors can be used
+#' as an edge weight.
+#' Javis and Patrick (1973) use a slightly
+#' modified (see parameter `jp`) shared nearest neighbor graph for
+#' clustering.
 #'
 #' @aliases sNN snn
 #' @family NN functions
@@ -43,10 +50,10 @@
 #' @param kt minimum threshold on the number of shared nearest neighbors to
 #' build the shared nearest neighbor graph. Edges are only preserved if
 #' `kt` or more neighbors are shared.
-#' @param jp use the definition by Javis and Patrick (1973), where shared
-#' neighbors are only counted between points that are in each other's
-#' neighborhood, otherwise 0 is returned. If `FALSE`, then the number of shared
-#' neighbors is returned, even if the points are not neighbors.
+#' @param jp In regular sNN graphs, two points that are not neighbors
+#' can have shared neighbors.
+#' Javis and Patrick (1973) requires the two points to be neighbors, otherwise
+#' the count is zeroed out. `TRUE` uses this behavior.
 #' @param search nearest neighbor search strategy (one of `"kdtree"`, `"linear"` or
 #' `"dist"`).
 #' @param sort sort by the number of shared nearest neighbors? Note that this
@@ -68,6 +75,7 @@
 #' \item{dist}{a matrix with the distances. }
 #' \item{shared }{a matrix with the number of shared nearest neighbors. }
 #' \item{k }{number of `k` used. }
+#' \item{metric }{the used distance metric. }
 #'
 #' @author Michael Hahsler
 #' @references R. A. Jarvis and E. A. Patrick. 1973. Clustering Using a
@@ -87,7 +95,7 @@
 #' # shared nearest neighbor distribution
 #' table(as.vector(nn$shared))
 #'
-#' # explore neighborhood of point 10
+#' # explore number of shared points for the k-neighborhood of point 10
 #' i <- 10
 #' nn$shared[i,]
 #'